third_party/blink/web_tests/fast/dom/shadow/layout-tests-can-access-shadow-expected.txt

CONSOLE WARNING: line 28: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
This tests that layout tests can access shadow DOM.


PASS shadow.nodeName is "#document-fragment"
PASS internals.shadowRoot(shadow) threw exception TypeError: Failed to execute 'shadowRoot' on 'Internals': parameter 1 is not of type 'Element'..
PASS internals.shadowRoot(p) is null
PASS shadow.nodeName is "#document-fragment"
PASS shadow === internals.shadowRoot(p) is true
PASS successfullyParsed is true

TEST COMPLETE
